Move OverviewInput from lambdaOverviews to datamodel
