--- app-backend/lambda-overviews/src/main/scala/com/rasterfoundry/lambda/overviews/OverviewInput.scala
@@ -1,24 +1,0 @@
-package com.rasterfoundry.lambda.overviews
-import java.util.UUID
-
-import io.circe.Decoder
-
-final case class OverviewInput(
-    outputLocation: String,
-    projectId: UUID,
-    projectLayerId: UUID,
-    refreshToken: String,
-    pixelSizeMeters: Int
-)
-
-object OverviewInput {
-
-  implicit val decodeOverview: Decoder[OverviewInput] =
-    Decoder.forProduct5("outputLocation",
-                        "projectId",
-                        "projectLayerId",
-                        "refreshToken",
-                        "pixelSizeMeters") {
-      OverviewInput.apply _
-    }
-}
